add header to primitives-traits
