+++ crates/primitives-traits/Cargo.toml
@@ -14,9 +14,14 @@ workspace = true
 [dependencies]
 reth-codecs.workspace = true
 
+alloy-consensus.workspace = true
+alloy-eips.workspace = true
 alloy-genesis.workspace = true
 alloy-primitives.workspace = true
-alloy-consensus.workspace = true
+alloy-rlp.workspace = true
+alloy-rpc-types.workspace = true
+
+revm-primitives.workspace = true
 
 # required by reth-codecs
 modular-bitfield.workspace = true
@@ -33,8 +38,10 @@ arbitrary = { workspace = true, features = ["derive"] }
 proptest.workspace = true
 proptest-derive.workspace = true
 test-fuzz.workspace = true
+rand.workspace = true
 
 [features]
+test-utils = ["arbitrary"]
 arbitrary = [
     "dep:arbitrary",
     "dep:proptest",
