+++ crates/primitives-traits/src/lib.rs
@@ -13,3 +13,9 @@
 /// Minimal account
 pub mod account;
 pub use account::Account;
+
+/// Common header types
+pub mod header;
+#[cfg(any(test, feature = "arbitrary", feature = "test-utils"))]
+pub use header::test_utils;
+pub use header::{Header, HeaderError, SealedHeader};
